DROTH-1536 (US 1492) Database and DAO: Alter table inaccurate_asset to accept link ids
