+++ digiroad2-oracle/src/main/scala/fi/liikennevirasto/digiroad2/dao/InaccurateAssetDAO.scala
@@ -13,8 +13,19 @@ class InaccurateAssetDAO {
       """.execute
   }
 
+  def createInaccurateLink(linkId: Long, typeId: Int, municipalityCode: Int, administrativeClass: AdministrativeClass) = {
+    sqlu"""
+        insert into inaccurate_asset (link_id, asset_type_id, municipality_code, administrative_class)
+        values ($linkId, $typeId, $municipalityCode, ${administrativeClass.value})
+      """.execute
+  }
+
   def getInaccurateAssetById(assetId: Long): Option[Long] = {
     sql"""select asset_id from inaccurate_asset where asset_id= $assetId""".as[Long].firstOption
+  }
+
+  def existsOnInaccurateWithLinkId(linkId: Long): Option[Long] = {
+    sql"""select linkId from inaccurate_asset where link_id= $linkId""".as[Long].firstOption
   }
 
   def getInaccurateAssetByTypeId(typeId: Int, municipalities: Set[Int] = Set(), adminClass: Set[AdministrativeClass] = Set()): List[(Long, String, Int)] = {
@@ -33,14 +44,34 @@ class InaccurateAssetDAO {
      """.as[(Long, String, Int)].list
   }
 
+  def getInaccurateByTypeId(typeId: Int, municipalities: Set[Int] = Set(), adminClass: Set[AdministrativeClass] = Set()): List[(Long, String, Int, Long)] = {
+
+    val withAuthorizedMunicipalities =
+      if (municipalities.nonEmpty) s" and ia.municipality_code in (${municipalities.mkString(",")})"  else s""
+
+    val withAdminClassRestrictions =
+      if(adminClass.nonEmpty) s" and ia.administrative_class in (${adminClass.map(_.value).mkString(",")})" else s""
+
+    sql"""
+       select ia.asset_id, m.name_fi, ia.administrative_class, ia.link_id
+       from inaccurate_asset ia
+       left join municipality m on ia.municipality_code = m.id
+       where ia.asset_type_id = $typeId #$withAuthorizedMunicipalities #$withAdminClassRestrictions
+     """.as[(Long, String, Int, Long)].list
+  }
+
+
   def deleteInaccurateAssetById(assetId: Long) = {
     sqlu"""delete from inaccurate_asset where asset_id= $assetId""".execute
+  }
+
+  def deleteInaccurateLinkId(linkId: Long) = {
+    sqlu"""delete from inaccurate_asset where link_id= $linkId""".execute
   }
 
   def deleteAllInaccurateAssets(typeId: Int) = {
     sqlu"""delete from inaccurate_asset
           where asset_type_id = $typeId""".execute
   }
-
 }
